test upgrade

+++ scripts/entrypoint.sh
@@ -2,29 +2,48 @@
 
 # exit script on any error
 set -e
-echo "setting up initial configurations"
+date=$(date '+%Y-%m-%d_%H:%M:%S')
+oldChainId="cosmoshub-3"
+GAIAD_HOME_COSMOSHUB_3_bak="${oldChainId}"-"${date}".bak
 
-if [ ! -f "$GAIAD_HOME/config/config.toml" ]; then
+chainIdInGenesis=$(cat -u  "${GAIAD_HOME}"/config/genesis.json 2>&1 | jq -r '.chain_id' 2>&1)
 
-	gaiad init "${MONIKER:-nonamenode}" --home="${GAIAD_HOME:-/.gaiad}" --chain-id="${CHAIN_ID:-cosmoshub-4}"
 
-	cd "$GAIAD_HOME/config"
+echo "chainIdInGenesis=$chainIdInGenesis,oldChainId=$oldChainId"
 
-	rm genesis.json
-	#rm genesis.cosmoshub-4.json
-	rm config.toml
-	rm app.toml
+if [ "$chainIdInGenesis" = "$oldChainId" ];then
+   echo "!!! backup $oldChainId home dir. !!!"
 
-	if [ ! -z "$GENESIS_URL" ]; then
-		wget "$GENESIS_URL"
-	else
-		wget https://github.com/cosmos/mainnet/raw/master/genesis.cosmoshub-4.json.gz
-		gzip -d genesis.cosmoshub-4.json.gz
-		cp genesis.cosmoshub-4.json genesis.json
-		rm genesis.cosmoshub-4.json
-	fi
+   echo "old dir"
+   ls
+   mv   "${GAIAD_HOME}"  "${GAIAD_HOME_COSMOSHUB_3_bak}"
+   echo "new dir"
+   ls
 
-	cat >config.toml <<EOF
+   echo "setting up initial configurations"
+   if [ ! -f "$GAIAD_HOME/config/config.toml" ]; then
+	 	gaiad init "${MONIKER:-nonamenode}" --home="${GAIAD_HOME}" --chain-id="${CHAIN_ID:-cosmoshub-4}"
+
+	 	mv  ${GAIAD_HOME}/config/priv_validator_key.json  ${GAIAD_HOME}/config/priv_validator_key.json-${date}_bak
+	 	cp  ${GAIAD_HOME_COSMOSHUB_3_bak}/config/priv_validator_key.json   ${GAIAD_HOME}/config/priv_validator_key.json
+
+	 	cd "$GAIAD_HOME/config"
+
+	 	rm genesis.json
+	 	#rm genesis.cosmoshub-4.json
+	 	rm config.toml
+	 	rm app.toml
+
+		if [ ! -z "$GENESIS_URL" ]; then
+			wget "$GENESIS_URL"
+		else
+			wget https://github.com/cosmos/mainnet/raw/master/genesis.cosmoshub-4.json.gz
+			gzip -d genesis.cosmoshub-4.json.gz
+			cp genesis.cosmoshub-4.json genesis.json
+			rm genesis.cosmoshub-4.json
+		fi
+
+		cat >config.toml <<EOF
 # This is a TOML config file.
 # For more information, see https://github.com/toml-lang/toml
 
@@ -562,6 +581,7 @@ EOF
 	#     lz4 -d -v --rm cosmoshub-3.20200415.0105.tar.lz4 | tar xf -
 	# fi
 
+fi
 fi
 
 exec supervisord --nodaemon --configuration /etc/supervisor/supervisord.conf
